+++ src/mjlab/tasks/velocity/config/booster_k1/rl_cfg.py
@@ -53,5 +53,5 @@ def booster_k1_ppo_runner_cfg() -> RslRlOnPolicyRunnerCfg:
     experiment_name="k1_velocity",
     save_interval=500,
     num_steps_per_env=24,
-    max_iterations=30_000,
+    max_iterations=15_000,
   )